requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
dj-database-url==2.3.0
Django==4.2.19
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==3.2.3
django-filter==24.3
django-redis==5.4.0
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework-camel-case==1.4.2
djangorestframework_simplejwt==5.4.0
drf-extra-fields==3.7.0
drf-yasg==1.21.8
factory_boy==3.3.3
Faker==36.1.1
filetype==1.2.0
gunicorn==23.0.0
inflection==0.5.1
kombu==5.4.2
packaging==24.2
pillow==11.1.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
